+++ tests/test_fallback_locale_as_callable.py
@@ -42,21 +42,32 @@ class TestClassic(Suite, ClassicTestCase):
     def create_tables(self):
         self.article = sa.Table(
             'article', self.metadata,
-            sa.Column('id', sa.Integer,
-                      autoincrement=True,
-                      primary_key=True,
-                      nullable=False),
+            sa.Column(
+                'id',
+                sa.Integer,
+                autoincrement=True,
+                primary_key=True,
+                nullable=False
+            ),
             sa.Column('locale', sa.types.CHAR(2), default='en'))
         self.article_translation = sa.Table(
             'article_translation', self.metadata,
-            sa.Column('id', sa.Integer, sa.ForeignKey('article'),
-                      primary_key=True,
-                      nullable=False),
-            sa.Column('locale', sa.types.CHAR(2),
-                      primary_key=True,
-                      nullable=False),
+            sa.Column(
+                'id',
+                sa.Integer,
+                sa.ForeignKey('article'),
+                primary_key=True,
+                nullable=False
+            ),
+            sa.Column(
+                'locale',
+                sa.types.CHAR(2),
+                primary_key=True,
+                nullable=False
+            ),
             sa.Column('name', sa.Unicode(255)),
-            sa.Column('content', sa.UnicodeText))
+            sa.Column('content', sa.UnicodeText)
+        )
 
     def create_models(self):
         class Article(ClassicBase, Translatable):